change: refactor global constant into the /components directory
