+++ frontend/components/footer.tsx
@@ -4,8 +4,8 @@ import {useUser} from "@/components/providers/user-provider";
 import {useEffect, useState} from "react";
 import {getClient} from "@/lib/graph/client";
 import {FooterSettingsDocument} from "@/lib/graph/generated/graphql";
-import {FOOTER_CONTACT_LINK_KEY, FOOTER_LEGAL_NOTICE_KEY} from "@/app/(settings)/app-settings/footer-form";
 import {useTranslations} from "next-intl";
+import {FOOTER_CONTACT_LINK_KEY, FOOTER_LEGAL_NOTICE_KEY} from "@/lib/constants/setting-keys";
 
 export function Footer() {
   const t = useTranslations("Components.Footer")
