+++ frontend/lib/constants/languages.ts
@@ -1,0 +1,10 @@
+export const AVAILABLE_LANGUAGES = [
+  {
+    name: "German",
+    localeKey: "de"
+  },
+  {
+    name: "English",
+    localeKey: "en"
+  }
+]
