+++ frontend/lib/constants/setting-keys.ts
@@ -1,0 +1,3 @@
+export const ABOUT_SECTION_TEXT_KEY = "ABOUT_SECTION_TEXT";
+export const FOOTER_LEGAL_NOTICE_KEY = "FOOTER_LEGAL_NOTICE_LINK"
+export const FOOTER_CONTACT_LINK_KEY = "FOOTER_CONTACT_LINK"
